# tests/unit_tests/fixtures/openapi.yaml
openapi: 3.0.0
info:
  title: V0 Semantic layer API
  version: 1.0.11
  contact:
    name: Андрей Абакин
    url: https://addressbook/search/profile/1763632
    email: abakin.a.v@sberbank.ru
  description: Semantic layer API for EPMPlatform
servers:
  - url: https://ci00163863-epmp-sndbx.apps.dev-terra000004-idm.ocp.delta.sbrf.ru
    description: DEV стенд
  - url: https://ci04763767-semlayer.apps.ift-terra000039-edb.ocp.delta.sbrf.ru
    description: IFT стенд
paths:
...
  /api/v0/tenants/:
    get:
      parameters:
        - example: 1
          name: page
          description: Страница (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationPageCount"
          in: query
          required: false
        - example: 20
          name: pageSize
          description: Количество строк (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationRowCount"
          in: query
          required: false
        - name: sortDirection
          description: Направление сортировки
          schema:
            $ref: "#/components/schemas/sortDirection"
          in: query
          required: false
      tags:
        - v0
      summary: Get all tenants
      operationId: getTenants
      responses:
        "200":
          description: Список всех тенантов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/tenant"
                maxItems: 1000
                uniqueItems: true
        "500":
          $ref: "#/components/responses/unexpectedError"
    post:
      tags:
        - v0
      summary: Create tenant
      operationId: createTenant
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tenant"
          description: Тенант создан
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Тенант уже существует
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/tenant"
        description: New tenant
  /api/v0/tenants/{tenantName}:
    get:
      tags:
        - v0
      summary: Get tenant by name
      operationId: getTenantByName
      responses:
        "200":
          description: Тенант найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tenant"
        "500":
          $ref: "#/components/responses/unexpectedError"
    delete:
      tags:
        - v0
      summary: Delete tenant by name
      operationId: deleteTenantByName
      responses:
        "200":
          description: Тенант `tenantName` удалена
        "404":
          description: Тенант `tenantName` не найден
        "500":
          $ref: "#/components/responses/unexpectedError"
    patch:
      tags:
        - v0
      summary: Change tenant by name
      operationId: changeTenantByName
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tenant"
          description: Тенант `tenantName` изменен
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Тенант `tenantName` не найден
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/tenantEdit"
    parameters:
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/search/objects:
    get:
      parameters:
        - example: 1
          name: page
          description: Страница (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationPageCount"
          in: query
          required: false
        - example: 20
          name: pageSize
          description: Количество строк (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationRowCount"
          in: query
          required: false
        - name: sortDirection
          description: Направление сортировки
          schema:
            $ref: "#/components/schemas/sortDirection"
          in: query
          required: false
        - name: query
          description: Поиск фрагмент технического имени или описания
          schema:
            $ref: "#/components/schemas/searchQuery"
          in: query
          required: True
        - name: type
          description: Тип объекта для фильтрации
          schema:
            $ref: "#/components/schemas/searchType"
          in: query
          required: false
        - name: modelName
          description: Имя модели для фильтрации
          schema:
            $ref: "#/components/schemas/modelName"
          in: query
          required: false
      tags:
        - v0
      summary: search for objects
      operationId: searchObjects
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/searchResponse"
                maxItems: 1000
                uniqueItems: true
          description: Результаты поиска
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Тенант или модель не найдена
        "500":
          $ref: "#/components/responses/unexpectedError"
    parameters:
      - $ref: "#/components/parameters/tenantName"
...
components:
  responses:
    unexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
  parameters:
...
    tenantName:
      name: tenantName
      description: Название тенанта
      schema:
        $ref: "#/components/schemas/tenantName"
      in: path
      required: true
...
  schemas:
    tenantName:
      $ref: "#/components/schemas/objectName32"
    dimensionName:
      $ref: "#/components/schemas/objectName32"
...
    compositeName:
      $ref: "#/components/schemas/objectName32"
    measureName:
      $ref: "#/components/schemas/objectName32"
    dataStorageName:
      $ref: "#/components/schemas/objectName64"
    modelName:
      $ref: "#/components/schemas/objectName32"
    objectName32:
      description: Уникальное техническое имя объекта
      type: string
      minLength: 1
      maxLength: 32
      pattern: "^[a-z0-9_]+$"
      example: "objectname32"
    objectName64:
      description: Уникальное техническое имя объекта
      type: string
      minLength: 1
      maxLength: 64
      pattern: "^[a-z0-9_]+$"
      example: "objectname64"
...
    tenantEdit:
      type: object
      additionalProperties: false
      properties:
        labels:
          $ref: "#/components/schemas/objectLabels"
        models:
          type: array
          items:
            $ref: "#/components/schemas/modelName"
          maxItems: 1000
          uniqueItems: true
    tenant:
      allOf:
        - $ref: "#/components/schemas/tenantEdit"
        - type: object
          required:
            - name
            - models
          additionalProperties: false
          properties:
            name:
              $ref: "#/components/schemas/tenantName"
...
    error:
      required:
        - message
      type: object
      additionalProperties: false
      properties:
        message:
          description: A human readable error message
          type: string
          pattern: "^[a-zA-Z0-9_]+$"
          minLength: 1
          maxLength: 255
...
    objectLabels:
      type: array
      maxItems: 100
      uniqueItems: true
      description: |
        Тексты, которые описывают различные сущности системы на разных языках для
          - признаков
          - атрибутов признаков
          - показателей
          - хранилищ данных
          - композитов
          - ...
      items:
        type: object
        additionalProperties: false
        required:
          - language
          - type
        description: Зависимый от языка текст различной длины
        properties:
          language:
            type: string
            maxLength: 5
            description: Язык текста
            example: ru-ru
            pattern: "^[a-z]{2}-[a-z]{2}$"
          type:
            $ref: "#/components/schemas/labelType"
          text:
            type: string
            description: Отображаемый текст
            pattern: '^[^\n\t]+$'
            minLength: 1
            maxLength: 255
...
    labelType:
      description: Вид метки для описания объекта - короткий для вывода в поле ограниченной длины, - длинный для отображения подробного описания
      type: string
      maxLength: 20
      enum:
        - SHORT
        - LONG
...
    searchQuery:
      description: Поиск фрагмента технического имени или описания
      type: string
      minLength: 1
      maxLength: 255
      pattern: '^[^\n\t]+$'
    searchType:
      description: Тип объекта фильтрации
      type: string
      maxLength: 20
      enum:
        - MEASURE
        - COMPOSITE
        - DIMENSION
        - DATA_STORAGE
...
    sortDirection:
      description: Направление сортировки
      default: asc
      enum:
        - asc
        - desc
      type: string
    paginationPageCount:
      example: 1
      description: Страница (параметр пагинации)
      type: integer
      minimum: 1
      maximum: 100
    paginationRowCount:
      example: 1
      description: Страница (параметр пагинации)
      type: integer
      minimum: 1
      maximum: 100
...
    searchResponse:
      type: object
      description: Результат поиска элементов по категориям
      additionalProperties: false
      properties:
        modelName:
          $ref: "#/components/schemas/SeacrhGroupItems"
...
    SeacrhGroupItems:
      additionalProperties: false
      description: Модель, содержащая результаты поиска
      type: object
      properties:
        dimensions:
          type: array
          description: Список измерений
          minItems: 0
          maxItems: 100
          items:
            $ref: "#/components/schemas/dimensionName"
        measures:
          type: array
          description: Список показателей
          minItems: 0
          maxItems: 100
          items:
            $ref: "#/components/schemas/measureName"
        dataStorages:
          type: array
          description: Список хранилищ данных
          minItems: 0
          maxItems: 100
          items:
            $ref: "#/components/schemas/dataStorageName"
        composites:
          type: array
          description: Список композитов
          minItems: 0
          maxItems: 100
          items:
            $ref: "#/components/schemas/compositeName"
...